+++ NEWS.md
@@ -1,4 +1,4 @@
-# tidygam 0.1.0.9000
+# tidygam 0.2.0
 
 ## Added